Mempercantik

+++ backend/notification-service/controllers/notificationController.js
@@ -4,36 +4,56 @@ const { publishNotification } = require('../configs/rabbidmq');
 const prisma = new PrismaClient();
 
 exports.getNotifications = async (req, res) => {
-    const notifications = await prisma.notification.findMany({
-        orderBy: { notifDate: 'desc' },
-    });
-    res.json(notifications);
+    // Get All
+    try {
+        const notifications = await prisma.notification.findMany({
+            orderBy: { notifDate: 'desc' },
+        });
+        res.json(notifications);
+    } catch (error) {
+        console.error("Error fetching notifications:", error);
+        res.status(500).json({ error: "Failed to retrieve notifications!" });
+    }
 };
 
 exports.deleteNotification = async (req, res) => {
+    // Parse
     const id = parseInt(req.params.id);
+    if (isNaN(id)) return res.status(400).json({ error: "ID invalid/tidak ada!" });
+    // Delete
     try {
         await prisma.notification.delete({ where: { notifId: id } });
         res.status(204).end();
     } catch (err) {
-        res.status(404).json({ error: 'Notification not found!' });
+        console.error("Error deleting notification:", error);
+        if (error.code === 'P2025') {
+            return res.status(404).json({ error: "Notification not found!" });
+        }
+        res.status(500).json({ error: 'Failed to delete notification!' });
     }
 }
 
 exports.postNotification = async (req, res) => {
+    // Parse
     const { itemId, message } = req.body;
-    
     if (!itemId || !message) {
-        return res.status(400).json({ error: "itemId dan message dibutuhkan!" });
+        return res.status(400).json({ error: "Satu atau beberapa field kosong!" });
     }
-    
-    // Save to DB
-    const newNotification = await prisma.notification.create({
-        data: { itemId, message },
-    });
-    
-    // Publish to RabbidMQ
-    await publishNotification({ itemId, message });
-
-    res.status(201).json(newNotification);
+    // Create
+    try {
+        // Save to DB
+        const newNotification = await prisma.notification.create({
+            data: {
+                itemId,
+                message
+            }
+        });
+        // Publish to RabbidMQ now!
+        await publishNotification({ itemId, message });
+        // Done
+        res.status(201).json(newNotification);
+    } catch (error) {
+        console.error("Error creating notification:", error);
+        res.status(500).json({ error: "Failed to create notification!" });
+    }
 }
